+++ apps/ipfs/connect
@@ -1,0 +1,7 @@
+#!/bin/bash
+
+export SCRIPT_DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" &> /dev/null && pwd )"
+export INPUT_PATH=$1
+
+echo "connecting to community $INPUT_PATH"
+IPFS_PATH=$INPUT_PATH ipfs daemon
