+++ apps/notes/.gitignore
